Ship the parsetab.py file in python distribution
The PLY parsing tables must be generated before we are able to start
parsing things. If they are not shipped in our distributions, they will
be generated during runtime. This may be an issue in environments where
the process running pyarn cannot create files in the same directory
where pyarn is installed.

Signed-off-by: Athos Ribeiro <athos@redhat.com>